KEEP-135 added changes based on refactoring suggestions
