+++ sdk/javascript/packages/azure/package.json
