+++ sdk/javascript/packages/azure/src/AzureSessionConfig.ts
@@ -1,0 +1,12 @@
+
+export class AzureSessionConfig {
+    tenantId: string;
+    clientId: string;
+    clientSecret: string;
+
+    constructor(tenantId: string, clientId: string, clientSecret: string) {
+        this.tenantId = tenantId;
+        this.clientId = clientId;
+        this.clientSecret = clientSecret;
+    }
+}
